pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snowpark-fundamentals"
version = "1.0.0"
description = "Snowpark ML Fundamentals - Week 1: Lunch & Learn Tutorial"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10,<3.13"
authors = [
    {name = "ML Engineering Team"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "snowflake-snowpark-python>=1.20.0",
    "snowflake-ml-python>=1.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "bandit>=1.7.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100
known_third_party = ["snowflake"]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "venv", ".venv", "build", "dist"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["venv", ".venv", "build", "dist"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.bandit]
skips = ["B101", "B105", "B608"]

[tool.coverage.run]
source = ["snowpark_fundamentals"]
omit = ["tests/*", "notebooks/*"]

[tool.coverage.report]
show_missing = true
fail_under = 70
